+++ src/main/java/act/cli/view/CliView.java
@@ -55,7 +55,7 @@ public enum CliView {
                 return;
             }
 
-            spec = PropertySpec.MetaInfo.withCurrent(spec, context);
+            spec = PropertySpec.MetaInfo.withCurrentNoConsume(spec, context);
 
             if (null == spec) {
                 spec = new PropertySpec.MetaInfo();
@@ -87,11 +87,11 @@ public enum CliView {
                 }
             }
             DataPropertyRepository repo = context.app().service(DataPropertyRepository.class);
-            List<String> outputFields = repo.outputFields(spec, componentType, context);
+            List<S.Pair> outputFields = repo.outputFields(spec, componentType, context);
             if (outputFields.isEmpty()) {
-                outputFields = C.list("this as Item");
+                outputFields = C.<S.Pair>list(S.pair("this as Item", "this as Item"));
             }
-            String tableString = cliContext.getTable(new CollectionASCIITableAware(dataList, outputFields, spec.labels(outputFields, context)));
+            String tableString = cliContext.getTable(new CollectionASCIITableAware(dataList, DataPropertyRepository.getFields(outputFields), spec.labels2(outputFields, context)));
             int itemsFound = dataList.size();
             CliCursor cursor = cliContext.session().cursor();
             String appendix = "";
@@ -238,12 +238,12 @@ public enum CliView {
             }
             componentType = firstElement.getClass();
             DataPropertyRepository repo = context.app().service(DataPropertyRepository.class);
-            spec = PropertySpec.MetaInfo.withCurrent(spec, context);
+            spec = PropertySpec.MetaInfo.withCurrentNoConsume(spec, context);
             if (null == spec) {
                 spec = new PropertySpec.MetaInfo();
                 spec.onValue("-not_exists");
             }
-            List<String> outputFields = repo.outputFields(spec, componentType, context);
+            List<S.Pair> outputFields = repo.outputFields(spec, componentType, context);
             if (outputFields.isEmpty()) {
                 return;
             }
@@ -257,9 +257,9 @@ public enum CliView {
             }
         }
 
-        private String buildDataLine(Object data, List<String> outputFields) {
-            Iterator<String> itr = outputFields.iterator();
-            String prop = itr.next();
+        private String buildDataLine(Object data, List<S.Pair> outputFields) {
+            Iterator<S.Pair> itr = outputFields.iterator();
+            S.Pair prop = itr.next();
             S.Buffer buf = S.buffer();
             buf.append(getProperty(data, prop));
             while (itr.hasNext()) {
@@ -268,22 +268,22 @@ public enum CliView {
             return buf.toString();
         }
 
-        private String getProperty(Object data, String prop) {
-            if ("this".equals(prop)) {
+        private String getProperty(Object data, S.Pair prop) {
+            if ("this".equals(prop._1)) {
                 return (escape(data));
             } else {
                 if (data instanceof AdaptiveRecord) {
-                    return escape(S.string(((AdaptiveRecord) data).getValue(prop)));
+                    return escape(S.string(((AdaptiveRecord) data).getValue(prop._1)));
                 }
-                return escape($.getProperty(data, prop));
+                return escape($.getProperty(data, prop._1));
             }
         }
 
-        private String buildHeaderLine(List<String> outputFields, Map<String, String> labels) {
+        private String buildHeaderLine(List<S.Pair> outputFields, Map<String, String> labels) {
             if (null == labels) {
                 labels = new HashMap<>();
             }
-            Iterator<String> itr = outputFields.iterator();
+            Iterator<S.Pair> itr = outputFields.iterator();
             String label = label(itr.next(), labels);
             S.Buffer buf = S.buffer();
             buf.append(label);
@@ -293,9 +293,13 @@ public enum CliView {
             return buf.toString();
         }
 
-        private String label(String key, Map<String, String> labels) {
-            String s = labels.get(key);
-            return null == s ? key : s;
+        private String label(S.Pair pair, Map<String, String> labels) {
+            String s = labels.get(pair._1);
+            String defLabel = pair._2;
+            if (null == defLabel) {
+                defLabel = pair._1;
+            }
+            return null == s ? defLabel : s;
         }
 
         private String escape(Object o) {