+++ src/main/java/act/conf/AppConfig.java
@@ -3714,4 +3714,8 @@ public class AppConfig<T extends AppConfig> extends Config<AppConfigKey> impleme
         }
         return timeStyle;
     }
+
+    public void setDefaultTldReloadCron() {
+        raw.put(canonical(TopLevelDomainList.CRON_TLD_RELOAD), "0 2 * * * *");
+    }
 }